+++ sidh/csidh/main.py
@@ -38,7 +38,7 @@ def csidh_main(ctx):
         # Maximum number of degree-(l_i) isogeny constructions is m (the same for each l_i)
         LABEL_m = 'with_same_bounds'
 
-    if setting.verbose:
+    if setting.tuned:
         verb = '-suitable'
     else:
         verb = '-classical'